Remove package-lock
Lockfiles are only respected for bare installs, not when installed as a
dep. Therefore, package locks are _primarily_ useful for apps, not libs.
Lockfiles are only useful for libs for locking down the dev environment.
Normally, this would be enough to warrant a lockfile. However, this
would mean that the lockfile would be used for CI as well. In this lib,
that means the nodenv version used for tests would be locked down.

So:
- the lockfile in this lib is only used for dev environment
- this package doesn't have that many devDeps
- the devDeps don't experience that much change
- we would prefer the nodenv devDep to always get latest on CI