Ensure required vars don't cause a failure exit
Exiting with a failure could cause npm to rollback the
install/uninstall and we don't want that.

Instead of erroring on missing vars, just emit a warning message.
Additionally, instead of defaulting to doing nothing in this error case,
we may as well still rehash. We could be doing more work than is
necessary, but it is an error case at that point after all. May as well
be maximally useful about it.