Only warn when _installing_ npm
technically, checking the lifecycle_event isn't necessary,
since if _uninstalling_ a buggy npm, the hook wouldn't be invoked
anyway. But it's nice to be explicit that we only want to show this
message when installing, not uninstalling, and only if the package is
npm, and only if that version is buggy.