post(un)install hook handles scopes and version specs
When a package is installed or uninstalled with a version spec (like
foo@latest or foo@1.2.3), the version spec is included in the argv blob.
So in order to match against the package name, we need to allow for the
version spec to exist (instead of matching up to the close bracket,
match to closing bracket _or_ @ sigil).